rename component files from `index` to their name
This makes for a better dev experience because not all files have the
same name, and each component's files are sorted the same way every
time (index is sorted after Button, but before Theme).
