+++ src/components/ThemePicker/ThemePicker.tsx
@@ -3,7 +3,7 @@ import LanguageContext from "../../contexts/LanguageContext";
 import { getActiveTheme, getTheme } from "../../utils/theme-utils";
 import { getTranslations } from "../../utils/translations-utils";
 import { Theme } from "../../types/Theme";
-import { ThemePickerTheme } from "../ThemePickerTheme";
+import { ThemePickerTheme } from "../ThemePickerTheme/ThemePickerTheme";
 import styles from "./ThemePicker.module.scss";
 
 type Props = {
